Create a shortcode for reusing content
This adds a new shortcode to allow for functionality similar to the Sphinx `.. include::` directive.
The idea is to have a content folder of partial/reusable content.
When called, the shortcode should pull the contents of the referenced markdown file, process it as normal, and add it to the rendered document in place of the shortcode.

Code is using a suggestion in a comment on this post: https://discourse.gohugo.io/t/best-practise-to-reuse-content/6855/2. 

This code would be called by something like
```
{{% include "path/to/file.md" %}}
```

Where the path should be something within the `content/` folder of doc site's repo.