Build warning fixes (#34)
In the upcoming Hugo 0.137.0 release, two params we use are being
removed having been deprecated in 0.120.0 and 0.124.0. This fixes those
params to the suggested updates. These deprecated params are causing
build errors on 0.136.1.

Hugo 0.129.0 updated **GitInfo** and requires that we add a new line
when setting up repos.
This modifies the `[Params]` section of `theme.toml` to include the new
required line.

---------

Co-authored-by: Rushan <rushenn@minio.io>